+++ src/lib.rs
@@ -38,17 +38,23 @@ use hyper_rustls::HttpsConnector;
 
 #[cfg(feature = "metrics")]
 use lazy_static::lazy_static;
-use opentelemetry::global;
-use opentelemetry::global::BoxedTracer;
-use opentelemetry::trace::Span;
-use opentelemetry::trace::Status;
 use quick_error::quick_error;
 use serde::{Deserialize, Serialize};
 use slog_scope::{error, info};
 use tokio::time::timeout;
 
+#[cfg(feature="opentelemetry")]
+use opentelemetry::global;
+#[cfg(feature="opentelemetry")]
+use opentelemetry::global::BoxedTracer;
+#[cfg(feature="opentelemetry")]
+use opentelemetry::trace::Span;
+#[cfg(feature="opentelemetry")]
+use opentelemetry::trace::Status;
+
 pub use types::*;
 
+#[cfg(feature="opentelemetry")]
 mod hyper_wrapper;
 /// The strongly typed data structures representing canonical consul objects.
 pub mod types;
@@ -217,6 +223,7 @@ impl Drop for Lock<'_> {
 pub struct Consul {
     https_client: hyper::Client<HttpsConnector<HttpConnector>, Body>,
     config: Config,
+    #[cfg(feature="opentelemetry")]
     tracer: BoxedTracer,
 }
 
@@ -226,7 +233,7 @@ fn https_client() -> HttpsConnector<HttpConnector> {
     #[cfg(feature = "rustls-webpki")]
     return hyper_rustls::HttpsConnectorBuilder::new().with_webpki_roots().https_or_http().enable_http1().build();
     #[cfg(feature = "default-tls")]
-    return HttpsConnector::new();
+    return hyper_tls::HttpsConnector::new();
 }
 
 impl Consul {
@@ -240,6 +247,7 @@ impl Consul {
         Consul {
             https_client,
             config,
+            #[cfg(feature="opentelemetry")]
             tracer: global::tracer("consul"),
         }
     }
@@ -677,6 +685,7 @@ impl Consul {
             )
             .body(body);
         let req = req.map_err(ConsulError::RequestError)?;
+        #[cfg(feature="opentelemetry")]
         let mut span = crate::hyper_wrapper::span_for_request(&self.tracer, &req);
 
         let method = req.method().clone();
@@ -704,6 +713,7 @@ impl Consul {
 
         let response = response?;
 
+        #[cfg(feature="opentelemetry")]
         crate::hyper_wrapper::annotate_span_for_response(&mut span, &response);
 
         let status = response.status();
@@ -731,7 +741,8 @@ impl Consul {
             Err(e) => {
                 record_failure_metric_if_enabled(&method, request_name);
 
-                span.set_status(Status::Error { description: e.to_string().into() });
+                #[cfg(feature="opentelemetry")]
+                span.set_status(Status::error(e.to_string()));
                 Err(ConsulError::InvalidResponse(e))
             }
         }