+++ CHANGELOG.md
@@ -8,6 +8,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 - `default-tls`, `rustls-native`, and `rustls-webpki` features to allow usage of rustls for the https client
 ### Changed
+- `opentelemetry` feature must be specified to enable opentelemetry support.
+- `opentelemetry` updated to version `0.18`.
+- `base64` updated to version `0.20`.
 ### Deprecated
 ### Removed
 ### Fixed