+++ CHANGELOG.md
@@ -6,6 +6,20 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## Unreleased
 
+## 0.5.0 - 2023-11-20
+
+### Added
+
+- Added support for `deregister-entity`.
+
+### Changed
+
+- `opentelemetry` updated to version `0.21` from `0.20`.
+
+### Changed
+
+- `opentelemetry` updated to version `0.20` from `0.19`.
+
 ## 0.4.0 - 2023-08-25
 
 ### Changed